chore: move saas config out of middleware
