+++ packages/workbench/src/ui/button.css.ts
@@ -35,10 +35,10 @@ export const secondary = stylesheet.style({
 
 export const small = stylesheet.style({
   fontSize: stylesheet.var('fontSizeXs'),
-  padding: '2px 2px',
+  padding: '4px 6px',
 })
 
 export const medium = stylesheet.style({
   fontSize: stylesheet.var('fontSizeSm'),
-  padding: '6px 6px',
+  padding: '8px 12px',
 })
